refactor : constants 파일명에 도메인 명시
