+++ src/service/ProfitService.js
@@ -1,4 +1,4 @@
-import { LOTTO_PRIZE, LOTTO_PRICE } from "../constants/systemConstants.js";
+import { LOTTO_PRIZE, LOTTO_PRICE } from "../lottoConstants/systemConstants.js";
 
 const calculateWinningAmount = (matchingCount) => {
   return Object.keys(matchingCount).reduce((sum, count) => sum + matchingCount[count] * (LOTTO_PRIZE[count] || 0), 0);
