+++ src/validation/validateBonusNumber.js
@@ -1,6 +1,6 @@
 import validationCondition from "./validateCondition.js";
 import runValidators from "../util/runValidators.js";
-import { BONUS_NUMBER_ERROR_MESSAGE } from "../constants/errorMessage.js";
+import { BONUS_NUMBER_ERROR_MESSAGE } from "../lottoConstants/errorMessage.js";
 
 const checkIsEmpty = (bonusNumberInput) => {
   if (validationCondition.isEmpty(bonusNumberInput)) {
